refactor: extract losses/rate_distortion.py
