+++ compressai/losses/__init__.py
@@ -27,75 +27,8 @@
 # OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 # ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from compressai import (
-    datasets,
-    entropy_models,
-    layers,
-    losses,
-    models,
-    ops,
-    transforms,
-    zoo,
-)
-
-try:
-    from .version import __version__
-except ImportError:
-    pass
-
-_entropy_coder = "ans"
-_available_entropy_coders = [_entropy_coder]
-
-try:
-    import range_coder
-
-    _available_entropy_coders.append("rangecoder")
-except ImportError:
-    pass
-
-
-def set_entropy_coder(entropy_coder):
-    """
-    Specifies the default entropy coder used to encode the bit-streams.
-
-    Use :mod:`available_entropy_coders` to list the possible values.
-
-    Args:
-        entropy_coder (string): Name of the entropy coder
-    """
-    global _entropy_coder
-    if entropy_coder not in _available_entropy_coders:
-        raise ValueError(
-            f'Invalid entropy coder "{entropy_coder}", choose from'
-            f'({", ".join(_available_entropy_coders)}).'
-        )
-    _entropy_coder = entropy_coder
-
-
-def get_entropy_coder():
-    """
-    Return the name of the default entropy coder used to encode the bit-streams.
-    """
-    return _entropy_coder
-
-
-def available_entropy_coders():
-    """
-    Return the list of available entropy coders.
-    """
-    return _available_entropy_coders
-
+from .rate_distortion import RateDistortionLoss
 
 __all__ = [
-    "datasets",
-    "entropy_models",
-    "layers",
-    "losses",
-    "models",
-    "ops",
-    "transforms",
-    "zoo",
-    "available_entropy_coders",
-    "get_entropy_coder",
-    "set_entropy_coder",
+    "RateDistortionLoss",
 ]
